requirements.txt
streamlit==1.32.0
snowflake-connector-python==3.7.0
snowflake-snowpark-python==1.12.0
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.30
mistralai==0.0.12
python-dotenv==1.0.1
speechrecognition==3.10.1
pydub==0.25.1
PyPDF2==3.0.1
streamlit-pdfviewer==0.0.3
streamlit-card-select==0.3.0
streamlit-card==0.0.4
streamlit-extras==0.3.6 